feat: 履歴一覧のリスト表示の実装 #11
